Extract file path helper partial
